Unify payload visitation into a single implementation by combining explicit roots and discovered system nexus roots
