--- temporalio/nexus/system/_payload_visitor.py
@@ -1,131 +1,0 @@
-from __future__ import annotations
-
-# This file is generated by gen_payload_visitor.py. Changes should be made there.
-from typing import Any
-
-import temporalio.nexus.system
-from temporalio.api.common.v1.message_pb2 import Payload
-from temporalio.bridge._visitor_functions import (
-    BoundedVisitorFunctions,
-    PayloadSequence,
-    VisitorFunctions,
-)
-
-
-class PayloadVisitor:
-    """A visitor for payloads.
-    Applies a function to every payload in a tree of messages.
-    """
-
-    def __init__(
-        self,
-        *,
-        skip_search_attributes: bool = False,
-        skip_headers: bool = False,
-        concurrency_limit: int = 1,
-    ):
-        """Creates a new payload visitor.
-
-        Args:
-            skip_search_attributes: If True, search attributes are not visited.
-            skip_headers: If True, headers are not visited.
-            concurrency_limit: Maximum number of payload visits that may run
-                concurrently during a single call to visit(). Defaults to 1
-                (sequential).
-        """
-        if concurrency_limit < 1:
-            raise ValueError("concurrency_limit must be positive")
-        self.skip_search_attributes = skip_search_attributes
-        self.skip_headers = skip_headers
-        self._concurrency_limit = concurrency_limit
-
-    async def visit(self, fs: VisitorFunctions, root: Any) -> None:
-        """Visits the given root message with the given function."""
-        method_name = "_visit_" + root.DESCRIPTOR.full_name.replace(".", "_")
-        method = getattr(self, method_name, None)
-        if method is None:
-            raise ValueError(f"Unknown root message type: {root.DESCRIPTOR.full_name}")
-        if self._concurrency_limit == 1:
-            await method(fs, root)
-            return
-
-        bounded = BoundedVisitorFunctions(fs, self._concurrency_limit)
-        try:
-            await method(bounded, root)
-        finally:
-            await bounded.drain()
-
-    async def _visit_nexus_operation_input_payload(
-        self,
-        fs: VisitorFunctions,
-        endpoint: str,
-        payload: Payload,
-    ) -> None:
-        new_payload = await temporalio.nexus.system._maybe_visit_payload(
-            endpoint,
-            payload,
-            fs,
-            self.skip_search_attributes,
-        )
-        if new_payload is None:
-            await self._visit_temporal_api_common_v1_Payload(fs, payload)
-            return
-
-        if new_payload is not payload:
-            payload.CopyFrom(new_payload)
-        await fs.visit_system_nexus_envelope(payload)
-
-    async def _visit_temporal_api_common_v1_Payload(
-        self, fs: VisitorFunctions, o: Payload
-    ):
-        await fs.visit_payload(o)
-
-    async def _visit_temporal_api_common_v1_Payloads(
-        self, fs: VisitorFunctions, o: Any
-    ):
-        await fs.visit_payloads(o.payloads)
-
-    async def _visit_payload_container(self, fs: VisitorFunctions, o: PayloadSequence):
-        await fs.visit_payloads(o)
-
-    async def _visit_temporal_api_common_v1_Memo(self, fs: VisitorFunctions, o: Any):
-        for v in o.fields.values():
-            await self._visit_temporal_api_common_v1_Payload(fs, v)
-
-    async def _visit_temporal_api_common_v1_SearchAttributes(
-        self, fs: VisitorFunctions, o: Any
-    ):
-        if self.skip_search_attributes:
-            return
-        for v in o.indexed_fields.values():
-            await self._visit_temporal_api_common_v1_Payload(fs, v)
-
-    async def _visit_temporal_api_common_v1_Header(self, fs: VisitorFunctions, o: Any):
-        for v in o.fields.values():
-            await self._visit_temporal_api_common_v1_Payload(fs, v)
-
-    async def _visit_temporal_api_sdk_v1_UserMetadata(
-        self, fs: VisitorFunctions, o: Any
-    ):
-        if o.HasField("summary"):
-            await self._visit_temporal_api_common_v1_Payload(fs, o.summary)
-        if o.HasField("details"):
-            await self._visit_temporal_api_common_v1_Payload(fs, o.details)
-
-    async def _visit_temporal_api_workflowservice_v1_SignalWithStartWorkflowExecutionRequest(
-        self, fs: VisitorFunctions, o: Any
-    ):
-        if o.HasField("input"):
-            await self._visit_temporal_api_common_v1_Payloads(fs, o.input)
-        if o.HasField("signal_input"):
-            await self._visit_temporal_api_common_v1_Payloads(fs, o.signal_input)
-        if o.HasField("memo"):
-            await self._visit_temporal_api_common_v1_Memo(fs, o.memo)
-        if o.HasField("search_attributes"):
-            await self._visit_temporal_api_common_v1_SearchAttributes(
-                fs, o.search_attributes
-            )
-        if o.HasField("header"):
-            await self._visit_temporal_api_common_v1_Header(fs, o.header)
-        if o.HasField("user_metadata"):
-            await self._visit_temporal_api_sdk_v1_UserMetadata(fs, o.user_metadata)
